Refactor tabs and enhance grid layout
