+++ tests/grid-tools.test.ts
@@ -14,13 +14,21 @@ describe('grid-tools', () => {
 
   test('applyGridLayout positions widgets', async () => {
     const items = [
-      { x: 0, y: 0, width: 10, height: 10, sync: jest.fn() },
-      { x: 0, y: 0, width: 10, height: 10, sync: jest.fn() },
+      { x: 0, y: 0, width: 10, height: 10, sync: jest.fn(), title: 'b' },
+      { x: 0, y: 0, width: 10, height: 10, sync: jest.fn(), title: 'a' },
     ];
-    const board = { selection: { get: jest.fn().mockResolvedValue(items) } };
-    await applyGridLayout({ cols: 1, rows: 2, padding: 5 }, board);
-    expect(items[1].y).toBe(15);
-    expect(items[1].x).toBe(0);
+    const board = {
+      selection: { get: jest.fn().mockResolvedValue(items) },
+      group: jest.fn(),
+    };
+    await applyGridLayout(
+      { cols: 1, rows: 2, padding: 5, sortByName: true, groupResult: true },
+      board,
+    );
+    // Items are sorted so 'a' is positioned first
+    expect(items[0].y).toBe(15);
+    expect(items[0].x).toBe(0);
+    expect(board.group).toHaveBeenCalledWith({ items: [items[1], items[0]] });
     expect(items[0].sync).toHaveBeenCalled();
     expect(items[1].sync).toHaveBeenCalled();
   });
